backend/requirements.txt
# --- FastAPI & Core Server ---
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0,<3.0.0
starlette==0.27.0
python-dotenv==1.0.0

# --- Real-time Communication ---
websockets==12.0

# --- Agentic AI & Orchestration ---
langchain>=0.1.0
langchain-groq>=0.0.1
langgraph>=0.0.10
langchain-core>=0.1.10
langchain-huggingface>=0.0.1
langchain_neo4j>=0.0.1

# --- Text Processing & Buffering ---
langchain-text-splitters>=0.0.1
sentence-transformers>=2.2.2

# --- Database Drivers ---
# Switched to latest for AsyncGraphDatabase support
neo4j>=5.15.0

# --- Networking & Utilities ---
httpx>=0.25.0
requests>=2.31.0
anyio>=3.7.0
typing-extensions>=4.15.0

neo4j 
pandas 
matplotlib 
textblob